refactor: change timespan to keystroke class
